Using localStorage instead of cookies for storing client-needed variables (username, max documents,...)
Previously I was using document.cookie to retrieve them, which is great
if all of your servers are one, but it's no longer a solution when your
API server is different from your React one.

There's a new Cookie Storage API which would give me the needed
flexibility, but it's still not fully supported by Firefox.

Note that localStorage is apparently restricted when private browsing in
iOS > 5, this might be troublesome